api/requirements.txt
fastapi
uvicorn
keras 
pandas 
numpy
sklearn.preprocessing
tensorflow
scikit-learn
yfinance